Split stack from current fiber
You can get stack from any fiber, not just current.
